updated axis from int to double

+++ Transformation/headers/Transformation.h
@@ -12,7 +12,7 @@ namespace Transform {
 		~Transformation();
 		void scale(Triangulation& inputTriangulation, Triangulation& outputTriangulation, double scalingFactor);
 		void processPoint(double value, Triangulation& outputTriangulation, std::map<double, int, Transformation>& uniqueValueMap, std::vector<int>& pointIndices);
-		void rotate(Triangulation& inputTriangulation, Triangulation& outputTriangulation, double angle, int axis[3]);
+		void rotate(Triangulation& inputTriangulation, Triangulation& outputTriangulation, double angle, double axis[3]);
 		void translate();
 		bool operator()(double a, double b) const;
 	};
